[ZF-9196] Unittests:
- Fixed detected if extension(pdo_sqlite) is loaded.

git-svn-id: http://framework.zend.com/svn/framework/standard/trunk@22569 44c647ce-9c0f-0410-b52a-842ac1e357ba

+++ tests/Zend/Db/Profiler/FirebugTest.php
@@ -106,7 +106,9 @@ class Zend_Db_Profiler_FirebugTest extends PHPUnit_Framework_TestCase
 
     public function tearDown()
     {
-        $this->_db->getConnection()->exec('DROP TABLE foo');
+        if (extension_loaded('pdo_sqlite')) {
+            $this->_db->getConnection()->exec('DROP TABLE foo');
+        }
 
         Zend_Wildfire_Channel_HttpHeaders::destroyInstance();
         Zend_Wildfire_Plugin_FirePhp::destroyInstance();
